Added File
This version creates tasks and then task by task creates python code.  It is not as good at this as I had hoped, probably because the individual tasks are created out of context of the others.  Version 11 will be attempting to do something similar with validation, but creating the code in a single retrieval.