Add file
This is version 11 that will try to do the same thing as 10 but creating the code in a single retrieval as is done in version 8, but with the validation as a single unit of work, which hopefully will prove more accurate, and also be substantially faster than 10.  The code as is does not yet work, but this is a starting point.