Update DSPY12.py
This is refactored to remove redundant code.  I have not had a chance to test it yet, as I am not at a computer where I can do so, but I think this should work, and is cleaner than the previous version.  Will test asap, and if necessary, update again with a final version.